Fixed project structure. All files that are only for the game engine are placed in engine folder. It now only has models services and thread sub-folders
