+++ 3DObjects/app/src/main/java/com/kamenov/martin/a3dobjects/command_promt_engine/contracts/Starter.java
@@ -1,0 +1,9 @@
+package com.kamenov.martin.a3dobjects.command_promt_engine.contracts;
+
+/**
+ * Created by Martin on 13.10.2018 г..
+ */
+
+public interface Starter {
+    void startEngine();
+}
