+++ 3DObjects/app/src/main/java/com/kamenov/martin/a3dobjects/constants/Constants.java
@@ -1,0 +1,28 @@
+package com.kamenov.martin.a3dobjects.constants;
+
+/**
+ * Created by Martin on 11.3.2018 г..
+ */
+
+public class Constants {
+    public static int SCREEN_WIDTH;
+    public static int SCREEN_HEIGHT;
+
+    public static String[] COLORS = {
+            "red",
+            "blue",
+            "green",
+            "white",
+            "black"
+    };
+
+    public static String[] COLORREPRESENTATIONS = {
+            "#ff0000",
+            "#0000ff",
+            "#00ff00",
+            "#ffffff",
+            "#000000"
+    };
+
+    public static float ROTATINGCOEF = 0.3f;
+}
