--- 3DObjects/app/src/main/java/com/kamenov/martin/a3dobjects/contracts/GameObject.java
@@ -1,11 +1,0 @@
-package com.kamenov.martin.a3dobjects.contracts;
-
-import android.graphics.Canvas;
-
-/**
- * Created by Martin on 6.3.2018 г..
- */
-
-public interface GameObject {
-    void update();
-}
