--- 3DObjects/app/src/main/java/com/kamenov/martin/a3dobjects/contracts/Rotatable.java
@@ -1,13 +1,0 @@
-package com.kamenov.martin.a3dobjects.contracts;
-
-/**
- * Created by Martin on 12.3.2018 г..
- */
-
-public interface Rotatable extends GameObject {
-    void rotateX3D(float delta);
-
-    void rotateY3D(float delta);
-
-    void rotateZ3D(float delta);
-}
