--- 3DObjects/app/src/main/java/com/kamenov/martin/a3dobjects/engine/contracts/Starter.java
@@ -1,9 +1,0 @@
-package com.kamenov.martin.a3dobjects.engine.contracts;
-
-/**
- * Created by Martin on 13.10.2018 г..
- */
-
-public interface Starter {
-    void startEngine();
-}
