+++ 3DObjects/app/src/main/java/com/kamenov/martin/a3dobjects/engine/models/game_objects/Sphere.java
@@ -1,0 +1,50 @@
+package com.kamenov.martin.a3dobjects.engine.models.game_objects;
+
+import android.graphics.Paint;
+
+import com.kamenov.martin.a3dobjects.engine.models.game_objects.contracts.DeepPoint;
+import com.kamenov.martin.a3dobjects.engine.models.game_objects.contracts.DrawingPart;
+import com.kamenov.martin.a3dobjects.engine.models.game_objects.contracts.Object3D;
+
+import java.util.ArrayList;
+
+/**
+ * Created by Martin on 21.10.2018 г..
+ */
+
+public class Sphere extends Object3D {
+    public float radius;
+    public Sphere(float x, float y, float z, Paint edgePaint, Paint wallPaint,
+                  float rotation, float radius) {
+        super(x, y, z, edgePaint, wallPaint, rotation);
+        this.radius = radius;
+        DeepPoint edgeAndWall = new DeepPoint(0, 0, 0);
+        points = new DeepPoint[] {edgeAndWall};
+
+        parts = new ArrayList<>();
+        parts.add(new DeepPoint[]{edgeAndWall});
+
+        setDrawingParts();
+    }
+
+    @Override
+    public void setDrawingParts() {
+        drawingParts = new ArrayList<>();
+        drawingParts.add(new DrawingPart(
+                x,
+                y,
+                z,
+                radius,
+                parts.get(0),
+                edgePaint,
+                getClass()));
+        drawingParts.add(new DrawingPart(
+                x,
+                y,
+                z,
+                radius,
+                parts.get(0),
+                wallPaint,
+                getClass()));
+    }
+}
