+++ 3DObjects/app/src/main/java/com/kamenov/martin/a3dobjects/engine/models/game_objects/contracts/Dimension.java
@@ -1,0 +1,11 @@
+package com.kamenov.martin.a3dobjects.engine.models.game_objects.contracts;
+
+/**
+ * Created by Martin on 25.12.2018 г..
+ */
+
+public enum Dimension {
+    X,
+    Y,
+    Z
+}
