+++ 3DObjects/app/src/main/java/com/kamenov/martin/a3dobjects/engine/models/game_objects/contracts/GameObject.java
@@ -1,0 +1,11 @@
+package com.kamenov.martin.a3dobjects.engine.models.game_objects.contracts;
+
+import android.graphics.Canvas;
+
+/**
+ * Created by Martin on 6.3.2018 г..
+ */
+
+public interface GameObject {
+    void update();
+}
