+++ 3DObjects/app/src/main/java/com/kamenov/martin/a3dobjects/engine/models/game_objects/contracts/MovingDirection.java
@@ -1,0 +1,14 @@
+package com.kamenov.martin.a3dobjects.engine.models.game_objects.contracts;
+
+/**
+ * Created by Martin on 12.11.2018 г..
+ */
+
+public enum MovingDirection {
+    Left,
+    Up,
+    Right,
+    Down,
+    Further,
+    Closer
+}
