--- 3DObjects/app/src/main/java/com/kamenov/martin/a3dobjects/models/game_objects_3d/contracts/MovingDirection.java
@@ -1,14 +1,0 @@
-package com.kamenov.martin.a3dobjects.models.game_objects_3d.contracts;
-
-/**
- * Created by Martin on 12.11.2018 г..
- */
-
-public enum MovingDirection {
-    Left,
-    Up,
-    Right,
-    Down,
-    Further,
-    Closer
-}
