--- 3DObjects/app/src/main/java/com/kamenov/martin/a3dobjects/models/services/contracts/Dimension.java
@@ -1,11 +1,0 @@
-package com.kamenov.martin.a3dobjects.models.services.contracts;
-
-/**
- * Created by Martin on 25.12.2018 г..
- */
-
-public enum Dimension {
-    X,
-    Y,
-    Z
-}
